test: rename files and clean-up
